requirements.txt
requests
pillow
imageio
ipython
requests
flask
python-dotenv